Introduce Commands, Tasks, Scheduler
This provides a basic infrastructure to run arbitrary Command
implementations on the core without blocking the coresession.

- Commands are created inside coresessions (always bound to a user ID)
- Tasks wrap commands for metadata (permissions, timings,..)
- Scheduler uses a QThreadPool to run the tasks (which are QRunnables)

By default, the _maximum_ amount of threads in the pool is the number of active
users at core start. Thereby every user _could_ be able to get a free
thread. There is no limitation on threads-per-user yet though.
Unused threads time out after 30 seconds.

The maximum threads can be overridden at compile time by defining CORE_MAX_TASK_THREADS.
Note: This can not be used to set the threads to zero. The minimum
number of threads is 1.

To enable the scheduler and thereby any kind of threads, use the core
command line switch --enable-tasks.